Remove redundat code

+++ LogicTest/LogicApiUT.cs
@@ -29,8 +29,8 @@ namespace LogicTest
             LogicApi logicApi = LogicApi.Instance();
             Assert.IsNotNull(logicApi);
             logicApi.CreateBalls(1);
-            float prev_x = logicApi.GetX(0);
-            float prev_y = logicApi.GetY(0);
+            double prev_x = logicApi.GetX(0);
+            double prev_y = logicApi.GetY(0);
             logicApi.LogicApiEvent += (sender, args) =>
             {
                 Assert.AreNotEqual(prev_y, logicApi.GetY(0));
